requirements.txt
aws-cdk-lib==2.160.0
constructs>=10.0.0
boto3
black==24.4.2
isort==5.13.2
moto
pytest==6.2.5